break up files, fix all type issues
