+++ lib/cache.ts
@@ -1,0 +1,4 @@
+import { Level } from "level"
+
+// Initialize LevelDB
+export const db = new Level("./pr-analysis-cache", { valueEncoding: "json" })
