+++ lib/getRepos.ts
@@ -1,4 +1,4 @@
-import { octokit } from "../index"
+import { octokit } from "./sdks"
 
 export async function getRepos(): Promise<string[]> {
   if (process.env.SHORT_REPO_LIST) {
@@ -29,7 +29,7 @@ export async function getRepos(): Promise<string[]> {
       page: page,
     })
 
-    repos = repos.concat(response.data.map((repo) => repo.full_name))
+    repos = repos.concat(response.data.map((repo: any) => repo.full_name))
 
     hasNextPage = response.data.length === 100
     page++
